+++ index.html
@@ -149,6 +149,6 @@
       <p id='historyStart'></p>
       <p id='subsStart'></p>
     </section>
-    <script src='../dist/script.js'></script>
+    <script src='dist/app/script.js'></script>
   </body>
 </html>
