+++ package.json
@@ -22,7 +22,7 @@
   "scripts": {
     "build": "tsc",
     "watch": "tsc -w",
-    "start": "npm run build && electron ./dist/main.js",
+    "start": "npm run build && electron ./dist/app/main.js",
     "package": "npm run build && electron-packager . --overwrite --out=releases --ignore='\\.idea'"
   },
   "main": "dist/main.js",
