+++ src/app/script.ts
