+++ widgets/timer.html
@@ -46,7 +46,6 @@
     </div>
     <script>
       // TODO move script to src/scripts or resources/scripts? typescript?
-      // TODO rename to timer.html
       // TODO store config instead of individual values?
       // timer
       let time = 0;
